Add first working prototype
Add the first working prototype of the Blazor Server side app.
This works and can do most things but is not by any means pretty
or good. There is NO regard for MVVM, and all classes interfaces
etc are a bit "hairy". But since it works it should now go into
SCM.