Conform to StyleCop and refactoring
Mostly formatting, readability and comments to conform to StyleCop
rules.
There is also some refactoring and syntax changes that improve
readability without changing what is happening ie.
Chores.cs line 179 which shortens the call to invoke cbChoreChanged.

Chore.cs also gets cleaned up for which fields are nullable as well
as changing the accessability for some properties. Since private
setters are now Deserializable from json it should be more strictly
managed.

The Change in Chore.Value to be nullable causes am exception in
ChoreMaker.razor, and need to be explicitly cast to decimal.