Lots of changes
Yes this is a terrible commit message.
Lots of things changed, but u can not be bothered to go into
everything and check for build status, and explain. Generally there
are also lots of CRLF issues, that makes this a giant mess. There
are also many changes to deal with nullablility and null handeling.

In broad terms this happend:
Counter.razor
- Is removed. That was just some sample that came with the Blazor
  template.

ChoreService.cs
- Lots of small changes to handle nullable fields and null checking
  where that is required.
- There was a problem where it would crash if there was no
  chores.json to deserialize. This is also a problem for worker.json
  but that is not fixed yet.
  It now makes a sample chore, serializes and exits to help get the
  application going from a fresh install.
- Many of the "special" get methods have been replaced with
  overloads of GetChoresAsync.

Program.cs
- Hardcodes a few paths. This is a very ugly quick fix to make this
  run as a windows service. It should obviousely be changed to
  get that out of appsettings.json